WIP: changes to raw_viewer that splits up the file into 2 classes, one for data manipulation and one for data visualization
